8298416: Console should be declared `sealed`
