+++ src/java.base/unix/native/libjava/Console_md.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2005, 2006, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2005, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -27,6 +27,7 @@
 #include "jni_util.h"
 #include "jvm.h"
 #include "java_io_Console.h"
+#include "java_io_ConsoleImpl.h"
 
 #include <stdlib.h>
 #include <unistd.h>
@@ -45,7 +46,7 @@ Java_java_io_Console_encoding(JNIEnv *env, jclass cls)
 }
 
 JNIEXPORT jboolean JNICALL
-Java_java_io_Console_echo(JNIEnv *env,
+Java_java_io_ConsoleImpl_echo(JNIEnv *env,
                           jclass cls,
                           jboolean on)
 {
